Boxplot: quantile whiskers
I added the keyword range_quantile. If set to true, the range value is not a multiple of the size of the box but a quantile. If set to 0.05, the outliers will be individuals below the 5th percentile and above the 95th percentile. 

I don't know if adding a kw was the best way but it allows to keep this non-breaking.